--- Deeploy/Targets/Snitch/TypeCheckers.py
@@ -1,53 +1,0 @@
-# ----------------------------------------------------------------------
-#
-# File: SnitchCheckers.py
-#
-# Last edited: 07.06.2024
-#
-# Copyright (C) 2024, ETH Zurich and University of Bologna.
-#
-# Author:
-# - Victor Jung, jungvi@iis.ee.ethz.ch, ETH Zurich
-#
-# ----------------------------------------------------------------------
-# SPDX-License-Identifier: Apache-2.0
-#
-# Licensed under the Apache License, Version 2.0 (the License); you may
-# not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an AS IS BASIS, WITHOUT
-# WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-from typing import List, Sequence, Type
-
-from Deeploy.AbstractDataTypes import Pointer
-from Deeploy.CommonExtensions.TypeCheckers.SignPropTypeChecker import SignPropTypeChecker
-from Deeploy.DeeployTypes import OperatorRepresentation, VariableBuffer
-
-
-class SnitchRQAddChecker(SignPropTypeChecker):
-
-    def __init__(self, input_types: Sequence[Type[Pointer]], output_types: Sequence[Type[Pointer]]):
-        super().__init__(input_types, output_types)
-
-    def _inferNumLevels(self, inputs: List[VariableBuffer],
-                        operatorRepresentation: OperatorRepresentation) -> List[int]:
-        return [operatorRepresentation['rqsOut_n_levels']]
-
-    def _inferSignedness(self, inputs: List[VariableBuffer],
-                         operatorRepresentation: OperatorRepresentation) -> List[bool]:
-        return [bool(operatorRepresentation["rqsOut_signed"])]
-
-    # Override this. This should compute the signednes of each output node of the Layer
-    def checkOutputType(self, inputs: List[VariableBuffer], operatorRepresentation: OperatorRepresentation) -> bool:
-        outputTypeSigned = self.output_types[0].referencedType.typeMin < 0
-        if operatorRepresentation['rqsOut_signed'] and outputTypeSigned:
-            return True
-        if (not operatorRepresentation['rqsOut_signed']) and (not outputTypeSigned):
-            return True
-        return False
